Done UI and offline rendering
Most of the UI feature are working and the possibility to view the
offline rendered images for the cam array was added.

Known issues:
	- viewing intersections doesnt work when a new camera was added
TODO:
	- grayscale view
	- view of all offline rendered images when on novel view,
	  currently segfaults
	- start working on reconstruction

+++ headers/commandManager.hpp
@@ -3,11 +3,12 @@
 #define GLFW_INCLUDE_VULKAN
 #include <GLFW/glfw3.h>
 
+// system includes
 #include <vector>
 #include <iostream>
 #include <array>
 
-extern const int MAX_FRAMES_IN_FLIGHT;
+extern const size_t MAX_FRAMES_IN_FLIGHT;
 struct QueueFamilyIndices;
 
 class CommandManager{
