+++ headers/syncManager.hpp
@@ -6,7 +6,7 @@
 #include <vector>
 #include <iostream>
 
-extern const int MAX_FRAMES_IN_FLIGHT;
+extern const size_t MAX_FRAMES_IN_FLIGHT;
 
 class SyncManager{
 public:
